services/asr/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxsentinel-asr"
version = "0.1.0"
description = "ASR Engine Abstraction Layer for VoxSentinel — routes audio to pluggable ASR backends"
requires-python = ">=3.11"
dependencies = [
    "tg-common",
    "deepgram-sdk>=3.0",
    "faster-whisper>=1.1",
    "websockets>=12.0",
    "circuitbreaker>=2.0",
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "numpy>=1.26",
    "torch>=2.3",
    "structlog>=24.2",
    "prometheus-client>=0.20",
    "redis>=5.0",
    "uvicorn>=0.30",
    "fastapi>=0.111",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["src/asr"]
